Ok, now it should be working....^^

+++ plot_g3d.py
@@ -592,9 +592,9 @@ class G3d_plot_line(G3d_plot):
 
         if self.args.range != None:
             ax.set_xlim(self.args.range)
-            idx = np.zeros(2)
-            idx[0] = (np.abs(self.x_array-self.args.range[0])).argmin()
-            idx[1] = (np.abs(self.x_array-self.args.range[1])).argmin()
+            idx = np.zeros(2,dtype=np.int8)
+            idx[0] = int((np.abs(self.x_array-self.args.range[0])).argmin())
+            idx[1] = int((np.abs(self.x_array-self.args.range[1])).argmin())
             idx = np.sort(idx)
             ax.set_ylim(np.min(self.line[idx[0]:idx[1]]))
